refactor: create mock data files for Card and MainPage tests
